User reported VPM kept creating old directory structure because of this. 
........
revert some minor stuff from last rev
........
Fix Champion Pub fast flip address
........
add Gilligan's Island L-8, thanks Inkochnito
........
fix tab order in paths dialog
........
Added LAH 1.06 Canadian, Dr. Dude LU-1, Elvira LF-1
........
Add fast flip support to Jack Bot
........
update
........
add VS2019 support
........
Fix the PREDCS_FIRQ_HACK for real
........
Improve HC55516 audio quality
........
Fix the PREDCS_FIRQ_HACK for real (missed file from commit 4736)
........
HC55516: fix missing clock stat collection for S11 games; add per-game option to initialize sample rate; add sample rate logging
........
Tweak to HC55516 rate logging to keep it in a separate file per ROM, and append...
........
Set initial hc55516 rates for Fliptronics 2 era games. 
........
A few more.
........
Made mingw compile possible again
........
Rolled back unneeded change
........
Added MC3417 chip (old HC55516 code) for Bally Vocalizer games
........
Whoops... removed a preview of coming attractions :)
........
Re-added leak / decay / charge values from exidy440, turned off ENABLE_LOWPASS_ESTIMATE
........
Using correct sound device interface (structure is the same as HC55516, that's why it worked)
........
Changing clock edge sensing as required per code comment
........
Set initial HC55516 clock rates for System 7 / 11 games
........
No fixed HC55516 clock rate for BK for now because french speech uses different clock
........
Added generic FIR lowpass filter when estimate filter is not used
........
Let's get the madness going: Start converting PinMAME over from th 'old MAME' license to the 3-Clause BSD License
........
move stone-old MAME files to subdir
........
add history.dat by Alexis Bousiges, which had a lot of nice content about the machines and manufacturers in it and is a bit hard to find nowadays
........
start moving some of the obvious files (i.e., PinMAME exclusive and with permission) to BSD3 license
........
delete again to try to restore history
........
add again with proper history
........
was rev 4745: Added MC3417 chip (old HC55516 code) for Bally Vocalizer games
........
was rev 4747: Re-added leak / decay / charge values from exidy440, turned off ENABLE_LOWPASS_ESTIMATE
........
was rev 4749: Changing clock edge sensing as required per code comment
........
was rev 4752: Added generic FIR lowpass filter when estimate filter is not used
........
start moving more of the obvious files (i.e., PinMAME exclusive and with permission) to BSD3 license
........
add a bit of info about the files
........
keep head mech reset hack to VPinMAME (also fixes compile issue on GCC)
........
hopefully fix GCC compile error
........
change default sample rate on unix also to 48khz and inc max to 96khz
........
really fix GCC compile issues
........
as the code in alsa.c and netbsd.c and arts.c somehow seems fishy (fractions are calculated in integer multiples of 22050 it seems?!?), set default sample rate to 44100 on unix instead of 48000
........
restore head mech handling also for the non-VPinMAME case (thanks DJRobX for the discussion on it) while (hopefully) fixing the GCC compile error
........
update
........
second round of HC55516 improvements, with automatic adaptation to variable sample rates and emulation of original hardware low-pass filters
........
add Road Show L-6C (competition MOD) by Soren
........
Update to LISY 5.15, by bontango. New: support for Bally/Stern 1977-85, incomplete WMS Sys7 support
........
update
........
fix hc55516 crash bug in F-14
........
explain the F-14 fix a little more thoroughly
........
delete lisy testprogs, by bontango
........
add Rollergames L-2C (competition MOD), and updates to Theatre of Magic 1.3XC (competition MOD); Mousin' Around! LA-4C (competition MOD); Creature from the Black Lagoon L-4C (competition MOD) by Soren
........
remove outdated info
........
move raspberry pi info
........
rename
........
update
........
add linux/ubuntu compile steps by legtod2
........
add some more info
........
Added POTO A3.1
........
Added Master by ManilaMatic, thanks to Stefan Fettes
........
Added 16-segments BCD conversion according to Rockwell 10939 / 10941 datasheets

git-svn-id: https://svn.code.sf.net/p/pinmame/code/branches/dll@4789 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/saa1099.c
@@ -342,7 +342,7 @@ int saa1099_sh_start(const struct MachineSound *msound)
 			name[j] = buf[j];
 			vol[j] = MIXER(intf->volume[i][j], j ? MIXER_PAN_RIGHT : MIXER_PAN_LEFT);
 		}
-		saa->stream = stream_init_multi(2, name, vol, sample_rate, i, saa1099_update);
+		saa->stream = stream_init_multi(2, name, vol, (int)sample_rate, i, saa1099_update);
 	}
 
 	return 0;
